+++ HISTORY.rst
@@ -3,6 +3,18 @@
 History
 -------
 
+0.1.3 (2020-02-26)
+---------------------
+
+* Adding credit card authorization capability
+
+0.1.2 (2020-02-25)
+---------------------
+
+* Adding transaction cancel, refund and rebill capabilities
+* Adding consult transaction information capability
+
+
 0.1.1 (2014-08-07)
 ---------------------